extracted all inventory filtering into its own component
